# Stellify-Update/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stellify Church Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /church/sermons:
    get:
      operationId: listSermons
      tags: [church-hub]
      summary: List biblical message recordings
      responses:
        "200":
          description: Sermons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sermon"
    post:
      operationId: createSermon
      tags: [church-hub]
      summary: Create a sermon entry (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SermonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sermon"
  /church/sermons/{id}:
    patch:
      operationId: updateSermon
      tags: [church-hub]
      summary: Update a sermon (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SermonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sermon"
    delete:
      operationId: deleteSermon
      tags: [church-hub]
      summary: Delete a sermon (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /church/events:
    get:
      operationId: listEvents
      tags: [church-hub]
      summary: List church events
      parameters:
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChurchEvent"
    post:
      operationId: createEvent
      tags: [church-hub]
      summary: Create a church event (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChurchEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChurchEvent"
  /church/events/{id}:
    patch:
      operationId: updateEvent
      tags: [church-hub]
      summary: Update an event (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChurchEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChurchEvent"
    delete:
      operationId: deleteEvent
      tags: [church-hub]
      summary: Delete an event (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /church/guides:
    get:
      operationId: listGuides
      tags: [church-hub]
      summary: List church guides and feeds
      responses:
        "200":
          description: Guides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guide"
    post:
      operationId: createGuide
      tags: [church-hub]
      summary: Create a guide (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guide"
  /church/guides/{id}:
    patch:
      operationId: updateGuide
      tags: [church-hub]
      summary: Update a guide (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuideUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guide"
    delete:
      operationId: deleteGuide
      tags: [church-hub]
      summary: Delete a guide (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ENGAGEMENT ───────────────────────────────────────────────────────────────
  /engagement/prayer-requests:
    get:
      operationId: listPrayerRequests
      tags: [engagement]
      summary: List prayer requests visible to current user
      responses:
        "200":
          description: Prayer requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrayerRequest"
    post:
      operationId: createPrayerRequest
      tags: [engagement]
      summary: Submit a prayer request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrayerRequestInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerRequest"
  /engagement/prayer-requests/{id}/complete:
    patch:
      operationId: completePrayerRequest
      tags: [engagement]
      summary: Mark a prayer request as done
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerRequest"
  /engagement/prayer-requests/{id}:
    patch:
      operationId: updatePrayerRequest
      tags: [engagement]
      summary: Member edits own pending prayer request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrayerRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerRequest"
  /engagement/counseling:
    get:
      operationId: listCounselingRequests
      tags: [engagement]
      summary: List counseling requests
      responses:
        "200":
          description: Counseling requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CounselingRequest"
    post:
      operationId: createCounselingRequest
      tags: [engagement]
      summary: Submit a counseling request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounselingRequestInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounselingRequest"
  /engagement/counseling/{id}/complete:
    patch:
      operationId: completeCounselingRequest
      tags: [engagement]
      summary: Mark counseling as done
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounselingRequest"
  /engagement/counseling/{id}:
    patch:
      operationId: updateCounselingRequest
      tags: [engagement]
      summary: Member edits own pending counseling request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounselingRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounselingRequest"
...
  /engagement/brethren-watch:
    post:
      operationId: createBrethrenWatch
      tags: [engagement]
      summary: Submit a Brethren Watch incident report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrethrenWatchInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrethrenWatch"
  /engagement/brethren-watch/{id}/resolve:
    patch:
      operationId: resolveBrethrenWatch
      tags: [engagement]
      summary: Resolve a Brethren Watch report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrethrenWatch"
...
components:
  schemas:
...
    Sermon:
      type: object
      required: [id, title, speaker, publishDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        speaker:
          type: string
        publishDate:
          type: string
        shortDescription:
          type: ["string", "null"]
        driveLink:
          type: ["string", "null"]
        createdAt:
          type: string
    SermonInput:
      type: object
      required: [title, speaker, publishDate]
      properties:
        title:
          type: string
        speaker:
          type: string
        publishDate:
          type: string
        shortDescription:
          type: ["string", "null"]
        driveLink:
          type: ["string", "null"]
    SermonUpdate:
      type: object
      properties:
        title:
          type: string
        speaker:
          type: string
        publishDate:
          type: string
        shortDescription:
          type: ["string", "null"]
        driveLink:
          type: ["string", "null"]
    ChurchEvent:
      type: object
      required: [id, title, eventDate, eventType, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        eventDate:
          type: string
        location:
          type: ["string", "null"]
        meetLink:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        eventType:
          type: string
          enum: [Worship, Caregroup, "Sharing Huddle", "Special Event", "On-site Event", "In Memory of the Lord Jesus", "Church Celebration"]
        createdAt:
          type: string
...
    PrayerRequestUpdate:
      type: object
      properties:
        description:
          type: string
        visibility:
          type: string
          enum: [Church, PastorOnly, ServantOnly, CaptainOnly]
...